web_version/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
